partly done . going to add phonenumber property to the register user dto and do some code

+++ Dto/Models/RegisterUserDto.cs
@@ -6,7 +6,7 @@ public class RegisterUserDto
     public string Email { get; set; }
     public string Password { get; set; }
     public int SexId { get; set; }
-    public int Age { get; set; }
+    public Int16 Age { get; set; }
     public int GradeId { get; set; }
     
 }
